Time: 14 ms (20.52%), Space: 23 MB (47.49%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
-    long long thours(vector<int>& arr, long long h) {
+    long long thours(vector<int>& arr, int h) {
         long long total = 0;
         for (int i = 0; i < arr.size(); i++) {
             total += ceil((double)arr[i] / h);
         }
         return total;
     }
-    int minEatingSpeed(vector<int>& arr, long long h) {
+    int minEatingSpeed(vector<int>& arr, int h) {
         int low = 1, high = *max_element(arr.begin(), arr.end());
 
         while (low <= high) {
